containers: Move tests to Compose Apps
