+++ factory-containers/test_tagging.py
@@ -1,10 +1,9 @@
 import json
-import os
-import subprocess
 import unittest
 
 from contextlib import contextmanager
 from tempfile import NamedTemporaryFile
+from target_manager import create_target as update_targets
 
 ONE_TO_ONE_JSON = {
     'targets': {
@@ -68,11 +67,7 @@ def temp_json_file(data):
 
 
 def create_target(tag, version, targets_file, apps):
-    args = ['./ota-dockerapp.py', 'create-target', version, targets_file]
-    args.extend(apps)
-    env = os.environ.copy()
-    env['OTA_LITE_TAG'] = tag
-    subprocess.check_call(args, env=env, cwd=os.path.dirname(__file__))
+    update_targets(targets_file, version, apps, tag, 'sadasd')
     with open(targets_file) as f:
         return json.load(f)
 
@@ -83,36 +78,36 @@ class TestTagging(unittest.TestCase):
         """Make sure the most basic one-to-one tagging method works."""
 
         with temp_json_file(ONE_TO_ONE_JSON) as filename:
-            create_target('devel', '13', filename, ['app1.dockerapp'])
+            create_target('devel', '13', filename, {'app1': {'uri': ''}})
             with open(filename) as f:
                 data = json.load(f)
                 target = data['targets']['raspberrypi-cm3-lmp-13']
                 # we should get the hash of the previous "devel" build
                 self.assertEqual('DEADBEEF', target['hashes']['sha256'])
                 self.assertEqual(
-                    ['app1'], list(target['custom']['docker_apps'].keys()))
+                    ['app1'], list(target['custom']['docker_compose_apps'].keys()))
 
                 target = data['targets']['minnow-lmp-13']
                 # we should get the hash of the previous "devel" build
                 self.assertEqual('ABCD', target['hashes']['sha256'])
                 self.assertEqual(
-                    ['app1'], list(target['custom']['docker_apps'].keys()))
+                    ['app1'], list(target['custom']['docker_compose_apps'].keys()))
 
         with temp_json_file(ONE_TO_ONE_JSON) as filename:
-            create_target('postmerge', '13', filename, ['app1.dockerapp'])
+            create_target('postmerge', '13', filename, {'app1': {'uri': ''}})
             with open(filename) as f:
                 data = json.load(f)
                 target = data['targets']['raspberrypi-cm3-lmp-13']
                 # we should get the hash of the previous "postmerge" build
                 self.assertEqual('G00DBEEF', target['hashes']['sha256'])
                 self.assertEqual(
-                    ['app1'], list(target['custom']['docker_apps'].keys()))
+                    ['app1'], list(target['custom']['docker_compose_apps'].keys()))
 
     def test_one_to_many(self):
         """Check that a single container build can create multiple Targets."""
 
         with temp_json_file(ONE_TO_MANY_JSON) as filename:
-            create_target('devel,postmerge', '13', filename, ['A.dockerapp'])
+            create_target('devel,postmerge', '13', filename, {'A': {'uri': ''}})
             with open(filename) as f:
                 data = json.load(f)
 
@@ -120,12 +115,12 @@ class TestTagging(unittest.TestCase):
                 target = data['targets']['raspberrypi-cm3-lmp-13-devel']
                 self.assertEqual('DEADBEEF', target['hashes']['sha256'])
                 self.assertEqual(
-                    ['A'], list(target['custom']['docker_apps'].keys()))
+                    ['A'], list(target['custom']['docker_compose_apps'].keys()))
 
                 target = data['targets']['minnow-lmp-13-devel']
                 self.assertEqual('ABCD', target['hashes']['sha256'])
                 self.assertEqual(
-                    ['A'], list(target['custom']['docker_apps'].keys()))
+                    ['A'], list(target['custom']['docker_compose_apps'].keys()))
                 self.assertEqual(['devel'], target['custom']['tags'])
 
                 # And one new target for "postmerge"
@@ -134,20 +129,20 @@ class TestTagging(unittest.TestCase):
                 self.assertEqual(['postmerge'], target['custom']['tags'])
                 self.assertEqual('G00DBEEF', target['hashes']['sha256'])
                 self.assertEqual(
-                    ['A'], list(target['custom']['docker_apps'].keys()))
+                    ['A'], list(target['custom']['docker_compose_apps'].keys()))
 
     def test_no_tags(self):
         """Make sure a factory with no tags defined works."""
 
         with temp_json_file(NO_TAGS_JSON) as filename:
-            create_target('', '13', filename, ['app1.dockerapp'])
+            create_target('', '13', filename, {'app1': {'uri': ''}})
             with open(filename) as f:
                 data = json.load(f)
                 target = data['targets']['raspberrypi-cm3-lmp-13']
                 # we should get the hash of the previous "devel" build
                 self.assertEqual('DEADBEEF', target['hashes']['sha256'])
                 self.assertEqual(
-                    ['app1'], list(target['custom']['docker_apps'].keys()))
+                    ['app1'], list(target['custom']['docker_compose_apps'].keys()))
 
     def test_derive_from_platform(self):
         """Assert that a tag like: postmerge-foo:postmerge works.
@@ -157,14 +152,14 @@ class TestTagging(unittest.TestCase):
 
         with temp_json_file(ONE_TO_ONE_JSON) as filename:
             create_target(
-                'postmerge-foo:postmerge', '13', filename, ['app1.dockerapp'])
+                'postmerge-foo:postmerge', '13', filename, {'app1': {'uri': ''}})
             with open(filename) as f:
                 data = json.load(f)
                 target = data['targets']['raspberrypi-cm3-lmp-13']
                 # we should get the hash of the previous "postmerge" build
                 self.assertEqual('G00DBEEF', target['hashes']['sha256'])
                 self.assertEqual(
-                    ['app1'], list(target['custom']['docker_apps'].keys()))
+                    ['app1'], list(target['custom']['docker_compose_apps'].keys()))
                 self.assertEqual(['postmerge-foo'], target['custom']['tags'])
